Se cambio el tipo de dato de la variable numero documento por Long y se agrego e implemento la foto del empleado

+++ src/main/java/com/proyecto/apartahotel/sispart/empleados/repository/IEmpleadoRepository.java
@@ -13,9 +13,9 @@ import com.proyecto.apartahotel.sispart.tipDocumento.entity.TipoDocumento;
 @Repository
 public interface IEmpleadoRepository extends JpaRepository<Empleado, Integer> {
 
-	Optional<Empleado> findByIdTipoDocumentoAndNumDocumento(TipoDocumento idTipoDocumento,int numDocumento);
+	Optional<Empleado> findByIdTipoDocumentoAndNumDocumento(TipoDocumento idTipoDocumento,Long numDocumento);
 
-	boolean existsByIdTipoDocumentoAndNumDocumento(TipoDocumento idTipoDocumento,int numDocumento);
+	boolean existsByIdTipoDocumentoAndNumDocumento(TipoDocumento idTipoDocumento,Long numDocumento);
 
 	void deleteByNumDocumento(int numDocumento);
